Restructure project to align with modern Rust conventions
- Move test-case from [dependencies] to [dev-dependencies] so it no
  longer ships in the runtime dependency graph.
- Replace wildcard re-exports in src/utils/http with explicit lists
  to make the public surface auditable.
- Tighten module visibility: clock and app are now private modules
  with selective pub use re-exports, matching the rest of core/infra.
- Extract long inline test modules (store, wheel, handlers,
  main_program, logger) to sibling tests.rs files via the
  submodule-as-file pattern, shrinking the source files while
  preserving private-item access.
- Convert all src/ mod.rs files to the file-as-module form
  (core.rs, infra.rs, utils.rs, utils/http.rs); tests/common/mod.rs
  retained as-is since integration test discovery requires it.
