+++ Cargo.toml
@@ -10,7 +10,6 @@ chrono = "0.4.44"
 ctrlc = "3.5.2"
 futures = "0.3.32"
 parking_lot = "0.12.5"
-test-case = "3.3.1"
 tokio = { version = "1", features = ["full"] }
 uuid = { version = "1.23.1", features = ["v4"] }
 hyper = { version = "1", features = ["full"] }
@@ -25,4 +24,5 @@ serde_json = "1.0.149"
 reqwest = { version = "0.13.3", features = ["json"] }
 
 [dev-dependencies]
+test-case = "3.3.1"
 tokio = { version = "1", features = ["test-util"] }
