+++ src/core.rs
@@ -3,7 +3,7 @@ mod store;
 mod timer;
 pub use timer::{TimeT, Timer, TimerId};
 
-pub mod clock;
+mod clock;
 pub use clock::{Clock, SystemClock};
 
 mod event_handler;
