+++ src/infra/handlers/tests.rs
@@ -1,0 +1,86 @@
+use crate::core::SystemClock;
+
+use super::*;
+use bytes::Bytes;
+use http::Method;
+use http_body_util::BodyExt;
+use hyper::body::Bytes as HyperBytes;
+use tokio::sync::mpsc;
+
+fn request(method: Method, body: &[u8]) -> HttpRequest {
+    HttpRequest {
+        method,
+        path: "/timer".to_string(),
+        body: HyperBytes::copy_from_slice(body),
+    }
+}
+
+async fn body_bytes(resp: HttpResponse) -> Bytes {
+    resp.into_body().collect().await.unwrap().to_bytes()
+}
+
+#[tokio::test]
+async fn status_handler_returns_ok() {
+    let handler = StatusHandler;
+    let req = HttpRequest {
+        method: Method::GET,
+        path: "/".to_string(),
+        body: HyperBytes::new(),
+    };
+
+    let resp = handler.handle(req).await.unwrap();
+    assert_eq!(resp.status(), StatusCode::OK);
+    assert_eq!(body_bytes(resp).await, Bytes::from("OK"));
+}
+
+#[tokio::test]
+async fn timer_handler_invalid_json_returns_bad_request() {
+    let clock = Arc::new(SystemClock);
+    let (sender, _receiver) = mpsc::channel::<TimerEvent>(1);
+    let handler = TimerHandler::new(sender, clock);
+
+    let resp = handler
+        .handle(request(Method::POST, b"not json"))
+        .await
+        .unwrap_err();
+
+    assert_eq!(resp.status(), StatusCode::BAD_REQUEST);
+    assert_eq!(body_bytes(resp).await, Bytes::from("Invalid JSON"));
+}
+
+#[tokio::test]
+async fn timer_handler_dropped_receiver_returns_service_unavailable() {
+    let clock = Arc::new(SystemClock);
+    let (sender, receiver) = mpsc::channel::<TimerEvent>(1);
+    drop(receiver);
+    let handler = TimerHandler::new(sender, clock);
+
+    let resp = handler
+        .handle(request(Method::POST, br#"{"interval_ms": 100}"#))
+        .await
+        .unwrap_err();
+
+    assert_eq!(resp.status(), StatusCode::SERVICE_UNAVAILABLE);
+    assert_eq!(body_bytes(resp).await, Bytes::from("Server overloaded"));
+}
+
+#[tokio::test]
+async fn timer_handler_happy_path_sends_event_and_returns_id() {
+    let clock = Arc::new(SystemClock);
+    let (sender, mut receiver) = mpsc::channel::<TimerEvent>(1);
+    let handler = TimerHandler::new(sender, clock);
+
+    let resp = handler
+        .handle(request(Method::POST, br#"{"interval_ms": 100}"#))
+        .await
+        .unwrap();
+
+    assert_eq!(resp.status(), StatusCode::OK);
+
+    let event = receiver.recv().await.unwrap();
+    let TimerEvent::Insert(timer) = event;
+
+    let body = body_bytes(resp).await;
+    let expected = format!("id: {}", timer.id.0);
+    assert_eq!(body, Bytes::from(expected));
+}
