--- src/utils/http/mod.rs
@@ -1,8 +1,0 @@
-mod router;
-pub use router::*;
-
-mod server;
-pub use server::*;
-
-mod utils;
-pub use utils::*;
